+++ packages/dappauth/rollup.config.js
@@ -1,14 +1,13 @@
-const commonjs = require('@rollup/plugin-commonjs');
-const resolve = require('@rollup/plugin-node-resolve');
-const json = require('rollup-plugin-json');
-const polyfills = require('rollup-plugin-polyfill-node');
-const copy = require('rollup-plugin-copy');
+import commonjs from '@rollup/plugin-commonjs';
+import resolve from '@rollup/plugin-node-resolve';
+import json from 'rollup-plugin-json';
+import polyfills from 'rollup-plugin-polyfill-node';
+import copy from 'rollup-plugin-copy';
 
-
-module.exports = [
+export default [
   // CommonJS
   {
-    input: 'index.js',
+    input: 'src/index.js',
     output: {
       file: 'dist/dappauth.js',
       format: 'cjs',
@@ -18,27 +17,25 @@ module.exports = [
     plugins: [
       copy({
         targets: [
-          { src: './ABIs', dest: 'dist' },
-          { src: './utils.js', dest: 'dist' },
+          { src: './src/ABIs', dest: 'dist' },
+          { src: './src/utils', dest: 'dist' },
         ],
       }),
+      commonjs(),
       resolve({
-        preferBuiltins: true,
         browser: true,
-      }),
-      commonjs({
-        include: 'node_modules/**',
+        preferBuiltins: true,
       }),
       json(),
       polyfills(),
     ],
     treeshake: {
       unknownGlobalSideEffects: false,
-    }
+    },
   },
   // es
   {
-    input: 'index.js',
+    input: 'src/index.js',
     output: {
       file: 'dist/dappauth.module.js',
       format: 'es',
@@ -47,16 +44,14 @@ module.exports = [
     plugins: [
       copy({
         targets: [
-          { src: './ABIs', dest: 'dist' },
-          { src: './utils.js', dest: 'dist' },
+          { src: './src/ABIs', dest: 'dist' },
+          { src: './src/utils', dest: 'dist' },
         ],
       }),
+      commonjs(),
       resolve({
-        preferBuiltins: true,
         browser: true,
-      }),
-      commonjs({
-        include: 'node_modules/**',
+        preferBuiltins: true,
       }),
       json(),
       polyfills(),
@@ -64,7 +59,7 @@ module.exports = [
   },
   // umd
   {
-    input: 'index.js',
+    input: 'src/index.js',
     output: {
       file: 'dist/dappauth.umd.js',
       format: 'umd',
@@ -73,16 +68,14 @@ module.exports = [
     plugins: [
       copy({
         targets: [
-          { src: './ABIs', dest: 'dist' },
-          { src: './utils.js', dest: 'dist' },
+          { src: './src/ABIs', dest: 'dist' },
+          { src: './src/utils', dest: 'dist' },
         ],
       }),
+      commonjs(),
       resolve({
         preferBuiltins: true,
         browser: true,
-      }),
-      commonjs({
-        include: 'node_modules/**',
       }),
       json(),
       polyfills(),