+++ packages/dappauth/src/ABIs/ERC1271.js
@@ -1,0 +1,25 @@
+export default [
+  {
+    constant: true,
+    inputs: [
+      {
+        name: 'hash',
+        type: 'bytes32',
+      },
+      {
+        name: '_signature',
+        type: 'bytes',
+      },
+    ],
+    name: 'isValidSignature',
+    outputs: [
+      {
+        name: 'magicValue',
+        type: 'bytes4',
+      },
+    ],
+    payable: false,
+    stateMutability: 'view',
+    type: 'function',
+  },
+];