+++ adapters/aptos-wallet-adapter-plugin/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @blocto/aptos-wallet-adapter-plugin
 
+## 0.2.7
+
+### Patch Changes
+
+- b9dac7d: support web3js v4
+- 8d0b5bf: enhance in-app-sdk event support
+- 00a3832: sendAsync can send another requests
+- Updated dependencies [b9dac7d]
+- Updated dependencies [8d0b5bf]
+- Updated dependencies [55afe21]
+- Updated dependencies [3017eb1]
+- Updated dependencies [00a3832]
+- Updated dependencies [aa28bf6]
+- Updated dependencies [d34fca2]
+  - @blocto/sdk@0.9.0
+
+## 0.2.7-beta.1
+
+### Patch Changes
+
+- b9dac7d: support web3js v4
+- 00a3832: sendAsync can send another requests
+- Updated dependencies [b9dac7d]
+- Updated dependencies [55afe21]
+- Updated dependencies [3017eb1]
+- Updated dependencies [00a3832]
+  - @blocto/sdk@0.9.0-beta.2
+
+## 0.2.7-beta.0
+
+### Patch Changes
+
+- 8d0b5bf: enhance in-app-sdk event support
+- Updated dependencies [8d0b5bf]
+  - @blocto/sdk@0.8.1-beta.0
+
+## 0.2.6
+
+### Patch Changes
+
+- Updated dependencies [a4d3fbd]
+  - @blocto/sdk@0.8.0
+
+## 0.2.6-beta.0
+
+### Patch Changes
+
+- Updated dependencies [4835522]
+- Updated dependencies [e742f66]
+  - @blocto/sdk@0.7.1-beta.0
+
 ## 0.2.5
 
 ### Patch Changes