+++ adapters/connectkit-connector/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @blocto/connectkit-connector
 
+## 0.2.2
+
+### Patch Changes
+
+- b9dac7d: support web3js v4
+- 8d0b5bf: enhance in-app-sdk event support
+- 00a3832: sendAsync can send another requests
+- Updated dependencies [b9dac7d]
+- Updated dependencies [8d0b5bf]
+- Updated dependencies [e075c22]
+- Updated dependencies [00a3832]
+  - @blocto/wagmi-connector@1.3.0
+
+## 0.2.2-beta.2
+
+### Patch Changes
+
+- Updated dependencies [fce0e50]
+  - @blocto/wagmi-connector@1.3.0-beta.2
+
+## 0.2.2-beta.1
+
+### Patch Changes
+
+- b9dac7d: support web3js v4
+- 00a3832: sendAsync can send another requests
+- Updated dependencies [b9dac7d]
+- Updated dependencies [00a3832]
+  - @blocto/wagmi-connector@1.2.4-beta.1
+
+## 0.2.2-beta.0
+
+### Patch Changes
+
+- 8d0b5bf: enhance in-app-sdk event support
+- Updated dependencies [8d0b5bf]
+  - @blocto/wagmi-connector@1.2.4-beta.0
+
 ## 0.2.1
 
 ### Patch Changes
@@ -7,6 +45,14 @@
 - 4835522: Fix export unreconize type for old typescript
 - Updated dependencies [4835522]
   - @blocto/wagmi-connector@1.2.2
+
+## 0.2.1-beta.0
+
+### Patch Changes
+
+- 4835522: Fix export unreconize type for old typescript
+- Updated dependencies [4835522]
+  - @blocto/wagmi-connector@1.2.2-beta.0
 
 ## 0.2.0