+++ packages/dappauth/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @blocto/dappauth
 
+## 2.2.1
+
+### Patch Changes
+
+- 9609a4e: fix: build config and test Switched from Mocha to Jest
+
+## 2.2.1-beta.0
+
+### Patch Changes
+
+- 9609a4e: fix: build config and test Switched from Mocha to Jest
+
 ## 2.2.0
 
 ### Minor Changes